requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-telegram-bot==20.7
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
apscheduler==3.10.4
pydantic-settings==2.1.0
pytest==8.0.0
httpx==0.27.0
